+++ semana18/LabECommerce/request.rest
@@ -1,10 +1,13 @@
-GET http://localhost:3003/user
+###
+GET http://localhost:3003/users
+
 
 ###
  POST http://localhost:3003/user
  Content-Type: application/json
  
  {
-     "name":
-
- }
+   "name":  "Magali",
+   "email": "maga@gmail.com",
+   "age": "22" 
+   }
