+++ semana18/LabECommerce/src/data/BaseDatabase.ts
@@ -10,7 +10,7 @@ export class BaseDatabase {
     client: "mysql",
     connection: {
        host: process.env.DB_HOST,
-       port: 3306,
+       port: 3303,
        user: process.env.DB_USER,
        password: process.env.DB_PASSWORD,
        database: process.env.DB_SCHEMA,
