+++ semana18/LabECommerce/src/data/ProductDatabase.ts
@@ -1,0 +1,24 @@
+import { BaseDatabase } from "./BaseDatabase";
+import { Product } from '../Classes/Product'
+
+export class ProductDataBase extends BaseDatabase {
+
+    public addProduct = async (product: Product) => {
+        await BaseDatabase.connection()
+            .insert(product)
+    }
+
+    public getAllProducts = async (): Promise<Product[]> => {
+        const result = await BaseDatabase.connection()
+            .select("*")
+        return result
+    }
+
+    
+    public getUserById = async (id:number): Promise<Product[]> =>{
+        const result = await BaseDatabase.connection("labEcommerce_user")
+        .select("*")
+        .where ("id", "=", `${id}`)
+        return result
+    }
+}
