+++ semana18/LabECommerce/src/data/connection.ts
@@ -7,7 +7,7 @@ export const connection = knex({
    client: "mysql",
    connection: {
       host: process.env.DB_HOST,
-      port: 3306,
+      port: 3303,
       user: process.env.DB_USER,
       password: process.env.DB_PASSWORD,
       database: process.env.DB_SCHEMA,
